requirements.txt
Django==2.2.1
djangorestframework==3.9.3
gunicorn==19.9.0
pytz==2019.1
sqlparse==0.3.0
xlrd==1.2.0
